change in predict and fast .py
