+++ detecting_fake_news/predict.py
@@ -1,64 +1,73 @@
 import joblib
 import pandas as pd
+import os
 
 from detecting_fake_news.preprocessing import TextPreprocessor
-from detecting_fake_news.params import BUCKET_NAME, PATH_TO_LOCAL_MODEL
-
-from google.cloud import storage
-import tensorflow as tf
+# from detecting_fake_news.params import BUCKET_NAME, PATH_TO_LOCAL_MODEL
+from detecting_fake_news.gcp import storage_download
 
 
-# run the predict from the local model.joblib
-def predict_local(text):
+#function that make the api home page, it is used in fast.py
+def home_page_api():
+    return {
+        '/test':
+        'For testing if the API is working, add 2 parameter. num1=NUMBER, num2=NUMBER',
+        '/predict_local':
+        'For running the API using the model stored localy, add 1 parameter. text=TEXT',
+        '/predict_cloud':
+        'For running the API using the model from the cloud, add 1 parameter. text=TEXT'
+    }
 
 
+#function that make the api test page, it is used in fast.py
+def tester_api(num1, num2):
+    return {'result':(int(num1) + int(num2))}
+
+
+#function that make the api prdict local page, it is used in fast.py
+def predict_local_api(text):
+    abspath = os.path.abspath("model.joblib")
+    # text form strign to df
     text = pd.DataFrame(dict(text=[text]))
-    print('working 1', text)
 
+    #clean text, remove punctuation, etc
     clean_text = TextPreprocessor().transform(text)
-    print('working 2', clean_text)
 
-    model = joblib.load('../model.joblib')
-    print('working 3', model)
+    #load the model.joblib that is already train
+    model = joblib.load(abspath)
 
+    # predict the model with the new text
     prediction_local = model.predict(clean_text)
-    print('working 4', prediction_local)
 
+    # change the output type
     results = int(prediction_local[0])
-    print('working 5')
 
-    return results
-
-
-
-#still in process
-# run the predict from the cloud model.joblib
-storage_client = storage.Client()
+    #api output
+    return {'prediction': results}
 
 
-#still in process
-def read_bucket_model(model_file_name):
-    gcp_model_path = f"gs://{BUCKET_NAME}/models/{model_file_name}"
 
-    loaded_model = joblib.load(tf.io.gfile.GFile(gcp_model_path, 'rb'))
-    return loaded_model
+#function that make the api predict cloud page, it is used in fast.py
+def predict_cloud_api(text):
+    abspath = os.path.abspath("cloud_model.joblib")
 
-
-#still in process
-def predict_cloud(text):
+    # text form strign to df
     text = pd.DataFrame(dict(text=[text]))
-    print('working 1', text)
 
+    #clean text, remove punctuation, etc
     clean_text = TextPreprocessor().transform(text)
-    print('working 2', clean_text)
 
-    model = joblib.load('../model.joblib')
-    print('working 3', model)
+    #load the model.joblib that is already train
 
-    prediction_local = read_bucket_model(PATH_TO_LOCAL_MODEL).predict(clean_text)
-    print('working 4', prediction_local)
+    storage_download('models/MultinomialNB/model.joblib', 'cloud_model.joblib')
 
+    model = joblib.load(abspath)
+
+    # predict the model with the new text
+    prediction_local = model.predict(clean_text)
+
+    # change the output type
     results = int(prediction_local[0])
-    print('working 5')
 
-    return results
+    #api output
+    return {'prediction': results}
